requirements.txt
asgiref==3.7.2
cffi==1.15.1
configparser==5.3.0
cryptography==41.0.1
Django==4.2.2
django-map-widgets==0.4.1
geographiclib==2.0
geopy==2.3.0
psycopg==3.1.9
psycopg2-binary==2.9.6
pycparser==2.21
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.7.0
